Add operation count comment generation, add header for function_definition
