+++ components/core/wf/code_generation/cpp_code_generator.cc
@@ -198,6 +198,13 @@ void cpp_code_generator::operator()(code_formatter& formatter, const ast::cast& 
                    make_view(x.arg));
 }
 
+void cpp_code_generator::operator()(wf::code_formatter& formatter, const ast::comment& x) const {
+  const std::vector<std::string> lines = x.split_lines();
+  for (const auto& line : lines) {
+    formatter.format("// {}\n", line);
+  }
+}
+
 void cpp_code_generator::operator()(code_formatter& formatter, const ast::compare& x) const {
   formatter.format("{} {} {}", make_view(x.left), string_from_relational_operation(x.operation),
                    make_view(x.right));
