+++ components/core/wf/code_generation/cpp_code_generator.h
@@ -28,6 +28,8 @@ class cpp_code_generator {
 
   void operator()(code_formatter& formatter, const ast::cast& x) const;
 
+  void operator()(code_formatter& formatter, const ast::comment& x) const;
+
   void operator()(code_formatter& formatter, const ast::compare& x) const;
 
   void operator()(code_formatter& formatter, const ast::construct_return_value& x) const;
